+++ src/pnfindiff/collocation.py
